feat(task selector): add selector logic (first iteration)

+++ task_selector/models.py
@@ -4,4 +4,4 @@ from pydantic import BaseModel
 class Task(BaseModel):
     name: str
     resources: List[str] = []
-    profit: int
+    profit: float
